pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stone_agent_core"
version = "0.1.0"
description = "A modular agent framework for building AI applications"
authors = [
    {name = "Wilfrido Jacobo Paredes", email = "wilfridojpg@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "langgraph>=0.6.8",
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "openai>=2.2.0",
    "anthropic>=0.69.0",
    "python-dotenv>=1.1.1",
]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["stone_agent_core"]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "black>=21.12b0",
    "isort>=5.10.1",
    "mypy>=0.910",
]